% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.R
\name{lcirc}
\alias{lcirc}
\title{Conduct LCIRC-N on a string.}
\usage{
lcirc(str, N = 1L)
}
\arguments{
\item{str}{one or more strings, as a character vector}

\item{N}{a single non-negative integer}
}
\value{
\code{str} after left circulating N times
}
\description{
Shift a string left N times, circulating overflow back onto the right side.
}
\examples{
lcirc("statistics", 3)
}
