% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.R
\name{rcirc}
\alias{rcirc}
\title{Conduct RCIRC-N on a string.}
\usage{
rcirc(str, N = 1)
}
\arguments{
\item{str}{a string, as a character vector}

\item{N}{a single non-negative integer}
}
\value{
\code{str} after right circulating N times
}
\description{
Shift a string right N times, circulating overflow back onto the left side.
}
\examples{
rcirc("statistics", 3)
}
